utils/requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.0
importlib-resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
rplidar==0.9.2
six==1.16.0
zipp==3.21.0
